# Conflicts:
#	openstudiocore/src/osversion/test/VersionTranslator_GTest.cpp
#	openstudiocore/src/utilities/CMakeLists.txt
#	openstudiocore/src/utilities/bcl/BCL.hpp
#	openstudiocore/src/utilities/bcl/RemoteBCL.cpp
#	openstudiocore/src/utilities/bcl/RemoteBCL.hpp
#	openstudiocore/src/utilities/core/String.cpp

+++ openstudiocore/src/model/SpaceLoadInstance.cpp
@@ -104,10 +104,10 @@ namespace detail {
     boost::optional<Space> space = this->space();
     boost::optional<SpaceType> spaceType = this->spaceType();
     if (space){
-      result = this->multiplier()*space->multiplier();
+      result = (int)(this->multiplier()*space->multiplier());
     }else if (spaceType){
       for (const Space& space : spaceType->spaces()){
-        result += this->multiplier()*space.multiplier();
+        result += (int)(this->multiplier()*space.multiplier());
       }
     }
